% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crepuscule.R
\name{crepuscule}
\alias{crepuscule}
\alias{crepuscule,sf,POSIXct-method}
\alias{crepuscule,matrix,POSIXct-method}
\alias{crepuscule,SpatialPoints,POSIXct-method}
\title{Compute crepuscular time}
\usage{
crepuscule(crds, dateTime, ...)

\S4method{crepuscule}{sf,POSIXct}(
  crds,
  dateTime,
  solarDep,
  direction = c("dawn", "dusk"),
  POSIXct.out = FALSE
)

\S4method{crepuscule}{matrix,POSIXct}(
  crds,
  dateTime,
  crs = sf::st_crs(4326),
  solarDep,
  direction = c("dawn", "dusk"),
  POSIXct.out = FALSE
)

\S4method{crepuscule}{SpatialPoints,POSIXct}(
  crds,
  dateTime,
  solarDep,
  direction = c("dawn", "dusk"),
  POSIXct.out = FALSE
)
}
\arguments{
\item{crds}{Geographical coordinates. It can be an object of
class \code{sf}, \code{matrix}, or \code{SpatialPoints}.}

\item{dateTime}{A \code{POSIXct} object representing the date and time. It specifies
the moment for which the crepuscular time is calculated.}

\item{...}{Additional arguments that are passed to methods.}

\item{solarDep}{A numerical value representing the solar depression angle.}

\item{direction}{A character string representing the \code{direction}, either "dawn" or "dusk".}

\item{POSIXct.out}{Logical, if TRUE, the result is returned as a \code{POSIXct} object,
otherwise, it is returned as a fraction of a day.}

\item{crs}{A \code{CRS} object representing the coordinate reference system.
Default is \code{sf::st_crs(4326)} which denotes WGS84 (World Geodetic System 1984).}
}
\value{
The function returns the time of crepuscular light, either as a fraction of a day
or as a \code{POSIXct} object, depending on the \code{POSIXct.out} parameter.
}
\description{
Calculates the crepuscular time, i.e., the time of dawn or dusk,
at a specific geographical location and time. Methods are available for different
object types with geographical coordinates, including:
\itemize{
\item \code{sf}: an object of class \code{sf}.
\item \code{matrix}: An unnamed matrix of coordinates, with each row containing a pair of geographical coordinates in \code{c(lon, lat)} order. See the example below.
\item \code{SpatialPoints}: an object of class \code{SpatialPoints}.
}
}
\details{
Input can consist of one location and at least one \code{POSIXct} time, or one \code{POSIXct} time and at least
one location. #' Do not use the daylight savings time zone string for supplying \code{dateTime}, as many OS will not be
able to properly set it to standard time when needed.
NOAA notes that “for latitudes greater than 72 degrees N and S, calculations are accurate to within
10 minutes. For latitudes less than +/- 72 degrees accuracy is approximately one minute.”
}
\examples{
#Civil dawn in Ithaca, NY on June 1, 2023

crepuscule(
       matrix(c(-76.4511, 42.4800), nrow = 1),
       as.POSIXct("2023-06-01", tz = "America/New_York"),
       solarDep = 6,
       direction = "dawn",
       POSIXct.out = TRUE
     )

}
