% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.R
\name{dls_correction}
\alias{dls_correction}
\title{Correct for daylight savings day length change}
\usage{
dls_correction(dateTime, lat)
}
\arguments{
\item{dateTime}{a POSIXct time.}

\item{lat}{a latitude in decimal degrees}
}
\value{
correction in seconds for longer/shorter daylength
}
\description{
helper function for identifying days and timezones affected by daylight savings time
}
\keyword{internal}
