% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun-methods.R
\name{solarnoon}
\alias{solarnoon}
\alias{solarnoon,sf,POSIXct-method}
\alias{solarnoon,matrix,POSIXct-method}
\alias{solarnoon,SpatialPoints,POSIXct-method}
\title{Compute solar noon time}
\usage{
solarnoon(crds, dateTime, ...)

\S4method{solarnoon}{sf,POSIXct}(crds, dateTime, POSIXct.out = FALSE)

\S4method{solarnoon}{matrix,POSIXct}(crds, dateTime, crs = sf::st_crs(4326), POSIXct.out = FALSE)

\S4method{solarnoon}{SpatialPoints,POSIXct}(crds, dateTime, POSIXct.out = FALSE)
}
\arguments{
\item{crds}{This represents geographical coordinates. It can be an object of
class \code{sf}, \code{matrix}, or \code{SpatialPoints}.}

\item{dateTime}{A \code{POSIXct} object representing the date and time. It specifies
the moment for which the solar noon is calculated.}

\item{...}{Additional arguments that are passed to methods.}

\item{POSIXct.out}{Logical, if \code{TRUE}, the result is returned as a \code{POSIXct} object, otherwise, it is returned as a fraction of a day.}

\item{crs}{A \code{CRS} object representing the coordinate reference system.
Default is \code{sf::st_crs(4326)}.}
}
\value{
The function returns the time of solar noon, either as a fraction of a day
or as a \code{POSIXct} object, depending on the \code{POSIXct.out} parameter.
}
\description{
Calculates the solar noon, i.e., the time when the sun is at its highest point in the sky.
at a specific geographical location and time. Methods are available for differet object types with
geographical coordinates, including:
\itemize{
\item \code{sf}: an object of class \code{sf}.
\item \code{matrix}: a matrix of coordinates.
\item \code{SpatialPoints}: an object of class \code{SpatialPoints}.
}
}
