% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunriset.R
\name{sunriset}
\alias{sunriset}
\alias{sunriset,sf,POSIXct-method}
\alias{sunriset,matrix,POSIXct-method}
\alias{sunriset,SpatialPoints,POSIXct-method}
\title{Calculate sunrise/sunset}
\usage{
sunriset(crds, dateTime, ...)

\S4method{sunriset}{sf,POSIXct}(
  crds,
  dateTime,
  direction = c("sunrise", "sunset"),
  POSIXct.out = FALSE
)

\S4method{sunriset}{matrix,POSIXct}(
  crds,
  dateTime,
  crs = sf::st_crs(4326),
  direction = c("sunrise", "sunset"),
  POSIXct.out = FALSE
)

\S4method{sunriset}{SpatialPoints,POSIXct}(
  crds,
  dateTime,
  direction = c("sunrise", "sunset"),
  POSIXct.out = FALSE
)
}
\arguments{
\item{crds}{Geographical coordinates. It can be an object of
class \code{sf}, \code{matrix}, or \code{SpatialPoints}.}

\item{dateTime}{A \code{POSIXct} object representing the date and time. It specifies
the moment for which the sunriset is calculated.}

\item{...}{Additional arguments that are passed to methods.}

\item{direction}{Character, determines whether to calculate the time of sunrise or sunset.}

\item{POSIXct.out}{Logical, if \code{TRUE}, the result is returned as a \code{POSIXct} object, otherwise, it is returned as a fraction of a day.}

\item{crs}{A "CRS" object representing the coordinate reference system.
Default is \code{sf::st_crs(4326)} which denotes WGS84 (World Geodetic System 1984).}
}
\value{
The function returns the time of sunriset, either as a fraction of a day
or as a \code{POSIXct} object, depending on the \code{POSIXct.out} parameter.
}
\description{
Calculates sunrise or sunset at a specific geographical location and time.
depending on the \code{direction} parameter. Methods are available
for different object types with geographical coordinates, including:
\itemize{
\item \code{sf}: an object of class \code{sf}.
\item \code{matrix}: An unnamed matrix of coordinates, with each row containing a pair of geographical coordinates in \code{c(lon, lat)} order. See the example below.
\item \code{SpatialPoints}: an object of class \code{SpatialPoints}.
}
}
\details{
Input can consist of one location and at least one \code{POSIXct} time, or one \code{POSIXct} time and at least
one location. \verb{solarDep`` is recycled as needed. Do not use the daylight savings time zone string for supplying }dateTime`, as many OS will not be
able to properly set it to standard time when needed.
}
\examples{
#Sunset in Ithaca, NY, USA on June 1, 2023

sunriset(
 matrix(c(-76.4511, 42.4800), nrow = 1),
 as.POSIXct("2023-06-01", tz = "America/New_York"),
 direction='sunset',
 POSIXct.out=TRUE
)
}
\references{
NOAA \href{https://gml.noaa.gov/grad/solcalc/sunrise.html}{sunrise/sunset calculator}
These algorithms include corrections for atmospheric refraction effects.
}
