% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun-methods.R
\name{crepuscule}
\alias{crepuscule}
\alias{crepuscule,sf,POSIXct-method}
\alias{crepuscule,matrix,POSIXct-method}
\alias{crepuscule,SpatialPoints,POSIXct-method}
\title{Compute crepuscular time}
\usage{
crepuscule(crds, dateTime, ...)

\S4method{crepuscule}{sf,POSIXct}(
  crds,
  dateTime,
  solarDep,
  direction = c("dawn", "dusk"),
  POSIXct.out = FALSE
)

\S4method{crepuscule}{matrix,POSIXct}(
  crds,
  dateTime,
  crs = sf::st_crs(4326),
  solarDep,
  direction = c("dawn", "dusk"),
  POSIXct.out = FALSE
)

\S4method{crepuscule}{SpatialPoints,POSIXct}(
  crds,
  dateTime,
  solarDep,
  direction = c("dawn", "dusk"),
  POSIXct.out = FALSE
)
}
\arguments{
\item{crds}{This represents geographical coordinates. It can be an object of
class "sf", "matrix", or "SpatialPoints".}

\item{dateTime}{A POSIXct object representing the date and time. It specifies
the moment for which the crepuscular time is calculated.}

\item{...}{Additional arguments that are passed to methods.}

\item{solarDep}{A numerical value representing the solar depression angle.}

\item{direction}{A character string representing the direction, either "dawn" or "dusk".}

\item{POSIXct.out}{Logical, if TRUE, the result is returned as a POSIXct object,
otherwise, it is returned as a fraction of a day.}

\item{crs}{A "CRS" object representing the coordinate reference system.
Default is sf::st_crs(4326).}
}
\value{
The function returns the time of crepuscular light, either as a fraction of a day
or as a POSIXct object, depending on the "POSIXct.out" parameter.
}
\description{
Calculates the crepuscular time, i.e., the time of dawn or dusk,
at a specific geographical location and time. Methods are available for different
object types with geographical coordinates, including:
\itemize{
\item "sf": an object of class "sf".
\item "matrix": a matrix of coordinates.
\item "SpatialPoints": an object of class "SpatialPoints".
}

Method for sf objects.

Method for matrix objects.

Method for SpatialPoints objects.
}
