% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun-methods.R
\name{solarnoon}
\alias{solarnoon}
\alias{solarnoon,sf,POSIXct-method}
\alias{solarnoon,matrix,POSIXct-method}
\alias{solarnoon,SpatialPoints,POSIXct-method}
\title{Compute solar noon time}
\usage{
solarnoon(crds, dateTime, ...)

\S4method{solarnoon}{sf,POSIXct}(crds, dateTime, POSIXct.out = FALSE)

\S4method{solarnoon}{matrix,POSIXct}(crds, dateTime, crs = sf::st_crs(4326), POSIXct.out = FALSE)

\S4method{solarnoon}{SpatialPoints,POSIXct}(crds, dateTime, POSIXct.out = FALSE)
}
\arguments{
\item{crds}{This represents geographical coordinates. It can be an object of
class "sf", "matrix", or "SpatialPoints".}

\item{dateTime}{A POSIXct object representing the date and time. It specifies
the moment for which the solar noon is calculated.}

\item{...}{Additional arguments that are passed to methods.}

\item{POSIXct.out}{Logical, if TRUE, the result is returned as a POSIXct object, otherwise, it is returned as a fraction of a day.}

\item{crs}{A "CRS" object representing the coordinate reference system.
Default is sf::st_crs(4326).}
}
\value{
The function returns the time of solar noon, either as a fraction of a day
or as a POSIXct object, depending on the "POSIXct.out" parameter.
}
\description{
Calculates the solar noon, i.e., the time when the sun is at its highest point in the sky.
at a specific geographical location and time. Methods are available for differet object types with
geographical coordinates, including:
\itemize{
\item "sf": an object of class "sf".
\item "matrix": a matrix of coordinates.
\item "SpatialPoints": an object of class "SpatialPoints".
}
}
