% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarpos.R
\name{solarpos}
\alias{solarpos}
\alias{solarpos,sf,POSIXct-method}
\alias{solarpos,matrix,POSIXct-method}
\alias{solarpos,SpatialPoints,POSIXct-method}
\title{Compute solar position}
\usage{
solarpos(crds, dateTime, ...)

\S4method{solarpos}{sf,POSIXct}(crds, dateTime, ...)

\S4method{solarpos}{matrix,POSIXct}(crds, dateTime, crs = sf::st_crs(4326), ...)

\S4method{solarpos}{SpatialPoints,POSIXct}(crds, dateTime, ...)
}
\arguments{
\item{crds}{Geographical coordinates. It can be an object of
class \code{sf}, \code{matrix}, or \code{SpatialPoints}.}

\item{dateTime}{A \code{POSIXct} object representing the date and time. It specifies
the moment for which the solar position is calculated.}

\item{...}{Additional arguments that are passed to methods.}

\item{crs}{A \code{CRS} object representing the coordinate reference system.
Default is \code{sf::st_crs(4326)}.}
}
\value{
Returns a matrix with the solar azimuth (in degrees from North), and elevation.
}
\description{
Calculates the solar position, i.e., the sun's elevation and azimuth, at a specific geographical location and time.
Methods are available for different object types with geographical coordinates, including:
\itemize{
\item \code{sf}: an object of class \code{sf}.
\item \code{matrix}: An unnamed matrix of coordinates, with each row containing a pair of geographical coordinates in \code{c(lon, lat)} order. See the example below.
\item \code{SpatialPoints}: an object of class \code{SpatialPoints}.
}
}
\details{
Input can consist of one location and at least one \code{POSIXct} time, or one \code{POSIXct} time and at least
one location. #' Do not use the daylight savings time zone string for supplying \code{dateTime}, as many OS will not be
able to properly set it to standard time when needed.
NOAA notes that “for latitudes greater than 72 degrees N and S, calculations are accurate to within
10 minutes. For latitudes less than +/- 72 degrees accuracy is approximately one minute.”
}
\examples{
# Solar position in Ithaca, NY, USA on June 1, 2023 at 08:00:00

solarpos(
 matrix(c(-76.4511, 42.4800), nrow = 1),
 as.POSIXct("2023-06-01 08:00:00", tz = "America/New_York")
)
}
\references{
#'NOAA \href{https://gml.noaa.gov/grad/solcalc/azel.html}{solar position calculator}
These algorithms include corrections for atmospheric refraction effects.
}
