% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun-methods.R
\name{solarpos}
\alias{solarpos}
\alias{solarpos,sf,POSIXct-method}
\alias{solarpos,matrix,POSIXct-method}
\alias{solarpos,SpatialPoints,POSIXct-method}
\title{Compute solar position}
\usage{
solarpos(crds, dateTime, ...)

\S4method{solarpos}{sf,POSIXct}(crds, dateTime, ...)

\S4method{solarpos}{matrix,POSIXct}(crds, dateTime, crs = sf::st_crs(4326), ...)

\S4method{solarpos}{SpatialPoints,POSIXct}(crds, dateTime, ...)
}
\arguments{
\item{crds}{This represents geographical coordinates. It can be an object of
class "sf", "matrix", or "SpatialPoints".}

\item{dateTime}{A POSIXct object representing the date and time. It specifies
the moment for which the solar position is calculated.}

\item{...}{Additional arguments that are passed to methods.}

\item{crs}{A "CRS" object representing the coordinate reference system.
Default is sf::st_crs(4326).}
}
\value{
The function returns a matrix with the sun's azimuth and elevation.
}
\description{
Calculates the solar position, i.e., the sun's elevation and azimuth, at a specific geographical location and time.
Methods are available for different object types with geographical coordinates, including:
\itemize{
\item "sf": an object of class "sf".
\item "matrix": a matrix of coordinates.
\item "SpatialPoints": an object of class "SpatialPoints".
}

Method for sf objects.

Method for matrix objects.

Method for SpatialPoints objects.
}
