% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarpos.R
\name{solarpos}
\alias{solarpos}
\alias{solarpos,sf,POSIXct-method}
\alias{solarpos,matrix,POSIXct-method}
\alias{solarpos,SpatialPoints,POSIXct-method}
\title{Compute solar position}
\usage{
solarpos(crds, dateTime, ...)

\S4method{solarpos}{sf,POSIXct}(crds, dateTime, ...)

\S4method{solarpos}{matrix,POSIXct}(crds, dateTime, crs = sf::st_crs(4326), ...)

\S4method{solarpos}{SpatialPoints,POSIXct}(crds, dateTime, ...)
}
\arguments{
\item{crds}{Geographical coordinates. It can be an object of
class \code{sf}, \code{matrix}, or \code{SpatialPoints}.}

\item{dateTime}{A \code{POSIXct} object representing the date and time.}

\item{...}{Additional arguments that are passed to methods.}

\item{crs}{A \code{CRS} object representing the coordinate reference system.
Default is \code{sf::st_crs(4326)}.}
}
\value{
Returns a matrix with the solar azimuth (in degrees from North), and elevation.
}
\description{
Calculates the solar position, i.e., the sun's elevation and azimuth, at a specific geographical location and time.
}
\details{
Methods are available for different classes of geographical coordinates, including:
\itemize{
\item \code{sf}: an object of class \code{sf}.
\item \code{matrix}: An unnamed matrix of coordinates, with each row containing a pair of geographical coordinates in \code{c(lon, lat)} order. See the example below.
\item \code{SpatialPoints}: an object of class \code{SpatialPoints}.
Input can consist of one location and at least one \code{POSIXct} time, or one \code{POSIXct}
time and at least one location. \code{solarDep} is recycled as needed.
Do not use the daylight savings time zone string for supplying \code{dateTime}, as many OS
will not be able to properly set it to standard time when needed.
}

Compared to NOAA’s original Javascript code, the sunrise and sunset estimates from this translation
may differ by +/- 1 minute, based on tests using selected locations spanning the globe.
This translation does not include calculation of prior or next sunrises/sunsets for locations above the Arctic
Circle or below the Antarctic Circle.

The calculations in the NOAA Sunrise/Sunset and Solar Position Calculators are based on equations
from Astronomical Algorithms, by Jean Meeus. The sunrise and sunset results are theoretically accurate
to within a minute for locations between +/- 72° latitude, and within 10 minutes outside of those latitudes.
However, due to variations in atmospheric composition, temperature, pressure and conditions, observed values may vary from calculations.

For the purposes of these calculators the current Gregorian calendar is extrapolated backward through time.
When using a date before 15 October, 1582, you will need to correct for this. The year preceding year 1 in
the calendar is year zero (0). The year before that is -1. The approximations used in these programs are very
good for years between 1800 and 2100. Results should still be sufficiently accurate for the range
from -1000 to 3000. Outside of this range, results may be given, but the potential for error is higher.

For sunrise and sunset calculations, we assume 0.833° of atmospheric refraction.
In the solar position calculator, atmospheric refraction is modeled as:

TODO, FILL OUT EQUATIONS AT https://gml.noaa.gov/grad/solcalc/calcdetails.html
}
\examples{
# Solar position in Ithaca, NY, USA on June 1, 2023 at 08:00:00

solarpos(
 matrix(c(-76.4511, 42.4800), nrow = 1),
 as.POSIXct("2023-06-01 08:00:00", tz = "America/New_York")
)
}
\references{
\itemize{
\item NOAA's \href{https://gml.noaa.gov/grad/solcalc/azel.html}{solar position calculator}
These algorithms include corrections for atmospheric refraction effects.
\item NOAA \href{https://gml.noaa.gov/grad/solcalc/calcdetails.html}{Solar Calculations Details}
}
}
