% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun-methods.R
\name{sunriset}
\alias{sunriset}
\alias{sunriset,sf,POSIXct-method}
\alias{sunriset,matrix,POSIXct-method}
\alias{sunriset,SpatialPoints,POSIXct-method}
\title{Calculate sunrise/sunset}
\usage{
sunriset(crds, dateTime, ...)

\S4method{sunriset}{sf,POSIXct}(
  crds,
  dateTime,
  direction = c("sunrise", "sunset"),
  POSIXct.out = FALSE
)

\S4method{sunriset}{matrix,POSIXct}(
  crds,
  dateTime,
  crs = sf::st_crs(4326),
  direction = c("sunrise", "sunset"),
  POSIXct.out = FALSE
)

\S4method{sunriset}{SpatialPoints,POSIXct}(
  crds,
  dateTime,
  direction = c("sunrise", "sunset"),
  POSIXct.out = FALSE
)
}
\arguments{
\item{crds}{This represents geographical coordinates. It can be an object of
class "sf", "matrix", or "SpatialPoints".}

\item{dateTime}{A POSIXct object representing the date and time. It specifies
the moment for which the sunriset is calculated.}

\item{direction}{Character, determines whether to calculate the time of sunrise or sunset.}

\item{POSIXct.out}{Logical, if TRUE, the result is returned as a POSIXct object, otherwise, it is returned as a fraction of a day.}

\item{crs}{A "CRS" object representing the coordinate reference system.
Default is sf::st_crs(4326).}
}
\value{
The function returns the time of sunriset, either as a fraction of a day
or as a POSIXct object, depending on the "POSIXct.out" parameter.
}
\description{
Calculates sunrise or sunset at a specific geographical location and time.
depending on the "direction" parameter. Methods are available
for differet object types with geographical coordinates, including:
\itemize{
\item "sf": an object of class "sf".
\item "matrix": a matrix of coordinates.
\item "SpatialPoints": an object of class "SpatialPoints".
}

Method for sf objects.

Method for matrix objects.

Method for SpatialPoints objects.
}
