% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suntools.R
\docType{package}
\name{suntools-package}
\alias{suntools}
\alias{suntools-package}
\title{Suntools: A package for calculating the sun's position relative to earth}
\description{
Sun and solar functions and methods formerly provided by the 'maptools' package.
}
\details{
Suntools is a package derived from the larger maptools package, now deprecated. It provides a set of functions
for calculating the sun's position, sunrise, sunset, polar noon, and crepuscular times
for any given geographical location and time on earth.

Major functions in the package include:
\itemize{
\item \code{crepuscule()}: Calculates crepuscular times, i.e. the time of dawn or dusk.
\item \code{sunrise()}: Calculates the times of sunrise and sunset at a given geographical location.
\item \code{solarpos()}: Calculates the solar position, including the sun's elevation and azimuth at a
specific geographical location and time.
\item \code{solarnoon()}: Calculates solar noon time at a given geographical location.
}

This package is particularly useful for studies related to solar energy,
environmental science, astronomy, and related fields.

For more details on specific functions, refer to their respective help pages.
}
\references{
The original maptools package: \url{https://CRAN.R-project.org/package=maptools}
}
\keyword{internal}
